test(mutation): add M15 and M16, and split the harness scaffolding out
M15 stops resolving markdown escapes and invisible characters; M16
collapses the regular-file tri-state back into a boolean. Both come back
DEAD. The arity guard moves 12 to 14.

Adding them took the module past 500 lines, so the scratch-worktree
machinery moves to _adr_debate_harness.py and the scratch_worktree
fixture to conftest.py. The fixture is in a conftest rather than imported
by name for a reason worth recording: an imported fixture reads as an
unused import, and `ruff check --fix` deleted it, turning all fourteen
mutants into setup errors. A conftest is the form a linter cannot remove.

The arity guard stays in the test module on purpose. It parses its own
__file__ for tests taking the fixture, so it has to sit beside the tests
it counts.

Also corrects two stale counts a review flagged: the inner suite is 70
cases across four files, not 62, and the inventory now lists M4 through
M16 rather than stopping at M12.

Refs #5205
